NCI-Agency/anet#3352: Reorganize and fix test cases
